requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
decorator==4.4.2
Django==3.1.2
future==0.18.2
geocoder==1.38.1
idna==2.10
Pillow==8.0.0
pytz==2020.1
ratelim==0.1.6
requests==2.24.0
six==1.15.0
asgiref==3.2.10
Django==3.1.2
djangorestframework==3.12.1
Pillow==8.0.0
sqlparse==0.4.1